Fix handling of -Y wrt name_arg count check
New jprint version at "0.0.20 2023-06-17".

Fix special handling for -Y option wrt exactly one name arg. The idea behind the
check is that if -G is specified then one cannot specify a name_arg. If it
is not specified then you must specify one name_arg. If both -G and a
name_arg are not specified it is also an error but there was a bug which
required that -G was used.