Add initial callback and calling code to jprint
Much more needs to be done here but the initial callback is added and
the code to call it is added and also called itself. Right now no
constraints on what to print (type, name, value, level etc.) have been
added and one might see both name and value printed. Not all types are
printed as it's not clear yet how to print all types. The functions are
definitely incomplete but this is a great start that will help when
actually implementing those constraints and then the right way to print
what is requested.