Add jprint -l level constraint check
More thorough testing must be done for this but initial support is at
least in.

I fixed some a wrong comment (yank paste error) and added a few notes
about the other number constraints and how it might have to be done (but
then it might not be done that way).

No patterns are looked for yet; it just goes through every pattern as
long as the depth is okay. It should be noted for patterns that this is
currently patterns requested rather than patterns found. That will have
to change and might involve a new list that is created before printing
the members. These are details TBD in a future update.