Add boolean to indicate if item is a value or name
Depending on whether the -Y option is specified or not we have to search
by value (default is to search by name) so we now have a boolean to
indicate if it's a value. This is not perfect and only will work (if
work it will) for very simple json files. It is currently unused in the
function that adds matches as both name and value have to be added to
each match. What will be important is whether to add the match to the
list based on whether it's a value or not.