Print name, value or both depending on options
If -Y is used by default it prints out the name of matches (though right
now no check of matches are done except for search by value or name i.e.
if it's not a value and we search by value it's not to be added). I am
not clear if this is correct. If it's supposed to print out the value as
well then it will have to be changed. But -p modifies this behaviour
anyway as noted below.

If -p both then print name and value. If -p v (default) then print only
value. If -p n print only the name. If one uses -Y they can still see
the name and value.

This has only been tested for very simple JSON files and the behaviour
in full might not be completely correct.