Partly implement pattern searching
Now prior to adding something to the matches list it will compare verify
that it's the same as the pattern requested. No regexp support yet.

Note that currently it only prints what is found and nothing else. It
does not consider if only value should be printed and if it's a value
for example. This can come later.