change dbg.clone and dyn_array.clone rules use ssh by default
The purpose of those rules is for someone who is maintaining
the repo to setup a proper remote clone using a GitHub
registered ssh key that will allow that person to push
changes into the other repo.

We add an echo to tell someone what to do if they don't
have a proper GitHub registered ssh key and thus the
clone rule will fail.  Such non-maintainers shouldn't
have a need to clone, but if they wished to, they can
simply enter the suggested git clone alternative
that will work without a Git registered ssh key.