New jprint version - fix level & name/value checks
New version is 0.0.22 2023-06-19.

There were two problems that were kind of connected that prevented
matches from being added to the matches list.

The jprint->search_value and is_value was indeed reversed which,
contrary to the previous comment, does indeed make more sense. The real
problem was that the level constraint was also buggy or more
specifically the depth being passed to the function was incorrect.

With this commit using -Y should only match values and using -l should
work. It should be noted however that this has only been tested for very
basic json files so the depth being passed might in fact be incorrect
and there might be cases where the is_value is also incorrect.