Implement jprint -L option
If -L is used then before printing the name and/or value of matches it
prints the level followed by the correct number of spaces or tabs
(spaces or tabs depending on the -L option argument) before printing the
name or value.

When -p both is used it prints each on a new line with the level and
correct number of spaces or tabs. This will most likely conflict with
some options later on probably including -j but whether this is an error
or not I do not yet know.

The way the spaces and tabs are printed is based on a loop. I could
swear there's another way to do it, a padding with an option arg to
printf() that says the number of tabs, but how to do this is escaping my
mind right now if it is indeed possible.