Implement jprint -B and -C (and partly -j)
As the options -B and -C are tied to -j I also in this commit have done
part of -j. For now it will only work with simple JSON files. If -B is
used and -j is also used then an opening '{' will be printed prior to
printing any matches and a closing '}' will be printed after the final
match. If no matches are specified this is not the case. If -C is
specified then after the final match is printed (or after the '}' if -B)
a final comma is printed.

A jprint_match now has a boolean string that if true means the match is
a string.

If -j specified then print the json basic syntax with the caveat that
values could end up being printed as name due to the way matching
currently is (it is not complete and what is there is not entirely
correct either).