Implement jprint -c (count only) option
Now if -c is used only matches of patterns found will be printed in the
same way that grep -c does. Note that without a patterns list there can
be no matches so the -c option will have no effect. If it's desirable to
have it print the number of each json member more work will have to be
done but this can be decided later on.

Since the options conflict quite a bit the use of -c with any of -B, -L,
-j and -I is an error except that if -I 0 is used it's not an error.
Perhaps it is an error and -I 0 should conflict with -c but I don't see
it as a problem either way.

Changed debug level 0 to debug level 1 as they've been in the way for
some time now that more is implemented.