Add boolean is_integer to struct json_number
Previously it only had booleans for floating point and exponents. More
correctly it had a boolean for each integer type but not a boolean to
simplify testing if it's a boolean or not. This will be useful for
jprint later on for -t and -Y processing.

New json parser and jparse versions for this: 1.0.6 2023-06-20.