Check jprint -t / -Y more for matches
Don't increment the count of matches if the type is not the same. By type this
refers to the `JTYPE_*` enum found in `json_parse.h`. This means the
`jprint_match` struct now has a type.

Add more strict checks for number types based on options.

Disable explicit check for strings in match function as the check there is
problematic anyway and this helps in another way until the function itself can
be fixed. Strings are more problematic than others but matching also
currently matches on either name or value when it should only match on
one or the other depending on the options specified.

Type checking is very much a work in progress.