jprint -l and -N cannot have max < 0
Only -n number ranges with max < 0 makes sense. -N especially does not
make sense. It is of questionable value whether or not -l should support
it but for now it does not. If it is desired to have it it can be added
another time.

Made sure to update test functions so that jprint -K does not fail so
that make test does not fail.

Verify that make prep works okay as this has failed a number of times
recently.