Implement jprint -L for -j
Now prints tabs or spaces after levels depending on the -L option when
printing syntax (-j). It already was done without -j. The text said that
it only has an effect with -j but I thought that it might be of use to
work without -j and that was done I think yesterday. If this is
undesired it can easily be fixed.