Implement jprint -s (substring matching)
What remains is regexp matching as far as matching goes. It should be
pointed out though that the matching is not complete even besides regexp
and some code will have to change in order to get matches to have values
that are different from the match itself.

These issues also hold for regular (exact) matching as well.