Implement jprint -I
This might have to change more when more of json syntax is printed (for
instance arrays) but for now this is done.

The way it works is a loop that prints the right number of tabs or
spaces based on the level or depth the matched node is in.

Note that without a match this will not happen. This must be corrected
later by adding every node as a match so that we don't print the json
document as a char * but rather by match like if a pattern is specified.