Fix jprint -s matching
Do not use strstr() or strcasestr() when finding previously added
matches as otherwise it will end up adding different matches as the
same. Put another way, only use the strstr()/strcasestr() functions in
matching itself not when finding previously added matches.

There still is the problem with the matched string being added as
the name and the value but that'll be worked out in the near future I
hope.