Implement jprint -P option and modularity
Implement jprint -P. When this option is disabled the spaces specified still
separate the tokens if printing both name and value. Default is 1 space.

When printing both (without json syntax) print on the same line the name and
value.

Added boolean to indicate that -b was used. Currently this is unused but it
will allow not relying on the spaces being not 0 as the default is 1
(this changed from 0 to 1 as it was always supposed to be 1 space by
default). This boolean is not yet needed but it might be and it's a
clear way to indicate if the option was used or not.

Make matching printing function a bit more modular by introducing a new
function, jprint_print_match(). Note that the concept of patterns is
incorrect and has to be fixed. Discussion is pending. Also note that
more functions will have to be added for printing matches.