Improve matching in jprint
This is still a work in progress but a bit of progress has been made
anyway.

In particular with substring matching:

    $ grep num h2g2.json
    "number":42,
    "number_as_str":"42",

    $ ./jprint -P -s -p both h2g2.json num
    num : number
    num : number_as_str

    $ ./jprint -P -s -p n h2g2.json num
    num
    num

    $ ./jprint -P -s -p v h2g2.json num
    number
    number_as_str

What this means is that what matched is not the same as the pattern
requested in all cases. Later it will be the actual name and/or value of
the json member but this is not currently possible. Note also that the
-Y option is not yet working properly which will affect this.

As well the struct jprint_match now has two struct json * nodes: the
name and value each of which depend upon if searching by value or name
was requested. Later both will be set (along with the above problem
being resolved) but for now only one will be non NULL. In any event
they're not used yet.