Fix display bug in vjson_fprint()
New jparse version 1.0.8 2023-06-23.

A problem occurred in vjson_fprint() where it would print the boolean
indicating if the number was sized as the specific type as the value
itself. This means that the number would not be printed but instead the
boolean would be printed cast to the type it is meant to indicate the
number is.