More jprint match printing modularity
Several new functions have been added to print parts of matches (or in
this case characters besides actual matches). In particular the
following functions were added:

    bool jprint_print_count(struct jprint *jprint);
    void jprint_print_final_comma(struct jprint *jprint);
    void jprint_print_brace(struct jprint *jprint, bool open);

These will be useful later for the printing of not just a JSON file if no
pattern requested (which currently is done by just printing the string
but because of levels and other things like constraints this will have
to be done differently) but also printing matches. More will be added
later.