Add initial (incomplete) man page for jprint tool
It is incomplete. In particular the options list is incomplete. The
synopsis is correct and the description is a good start (but needs some
work). Examples section is a good start but the commands don't show
correct output yet either except in one condition. Bugs section does not
list these problems as the tool is not finished yet.

I wanted the man page in the repo and it can be discussed and I can also
finish adding the options hopefully in the coming days.

jparse/Makefile updated so that make install will install jprint.1 along
with the other man pages.