Add option -Y to man page
Why is this option -Y?  Why not Y?  Because Y, that's why Y!  Why
besides, all the other good options were already taken. :-) Specifically
the letter Y has a V in it and V would have been the obvious choice but
both v and V are taken.  This is why you'll have to believe us when we
tell you that this is a good enough reason why Y! :-)