Only have -j -p an error if not -p b or -p both
New jprint version 0.0.27 2023-06-24.

If -j is used don't make use of -p b or -p both an error. It's only an
error if printing of just name or just value is specified (after the -j
as -j will set both). Checking for this is just as simple as for -p
being used at all and it seems slightly more user-friendly to do it this
way.