Make jprint options/file checking more modular
In other words jprint_sanity_chks() does a lot more work now. It checks
options combinations, verifies that a path is specified and it exists as
a regular readable file. If this is not true the function will not
return (it will not return on any other error either). It will never
return a NULL pointer (it returns a FILE *) but we check for a NULL
pointer in main() anyway for extra defence.

It does NOT look at name_args and it does NOT verify it is valid JSON.
Currently that is main()'s job but these things might be moved to other
functions as well.

As the argc and argv need to be shifted inside main() the
jprint_sanity_chks() function takes a int *argc and char ***argv rather
than the usual int argc and char **argv.