Add parse_jprint_name_args() for more modularity
Move the searching on the command line for name_args to a separate
function (was in main()) which the jprint_sanity_chks function calls.
The reason jprint_sanity_chks() calls it is because some options have to
be checked after name_args are found: -Y requires exactly one name_arg
(!= 1 is an error) and -c cannot be used if a name_arg is specified. As
well no name_arg can be specified if we're printing the entire file
(this is probably an unnecessary check as an artefact from the removed
-o option but this does not hurt either).