Fix bug with -S and make it more modular
The code that was in main() that runs the JSON check tool if -S path
used is now in a new function, run_jprint_check_tool(). The stream for
the pipe is now in the struct jprint. The file contents, the json.file
FILE *, the tool path and tool args are also all in the struct jprint.

This fixes a bug too where the file.json could be printed more than once
if -S was used. For instance doing:

    jprint -S /bin/cat h2g2.json

used to print the file three times but now it prints it just once which
I believe is the correct behaviour.