Remove jprint -G option
Although I originally had some use cases it was brought up that it might
be not that useful and with the example files that we have nothing is
coming to mind that might make it useful so I have removed it.

It should be noted that the jprint_pattern struct (which is not
completely correct in concept but it will still exist in some form or
another, whether renamed or not, when this is corrected) still has the
boolean use_regexp (just as it has use_substrings and use_value) to
indicate what kind of search the pattern is (even though it's probably
now okay to have them just in the struct jprint it doesn't hurt to have
it in the other struct and they might have some use that is not yet
shown).