Modularise jprint vjprint_json_search()
By a new function is_jprint_match() that takes the jprint struct, the
pattern, the json node and the string that matched. This simplifies the
switch block a fair bit where the new function only checks for a match,
returning true or false, and the other one calls that function and only
if it does match (true) does it attempt to add it to the matches list.

I've become aware of another bug that might be worth pointing out that
has to do with processing of patterns: if doing:

    ./jprint -Y int -j h2g2.json 42

one will not get only the two 42 integers but also the string "42" (of
course currently the value or name is incorrect, being the same as what
matched but that's something that will be worked out later too). I do
have some thoughts on this but this is again very much a work in
progress and right now I am mostly focusing on modularity to help with
other things.

This issue has been noted in a comment as well.