Clarify error of use of both -p and -j in usage
As of yesterday if one does -p both / -p b it's okay to use -j as -j
sets -p both. This is just as easy to check as it is to check for
explicit use of -p full stop and it's slightly more user friendly so I
made that change but I didn't fix it in the usage string.