Move most code from jprint.c to jprint_util.c
Since most of the functions in jprint.c are utility in nature I have
moved them to jprint_util.c. This required moving the structs that were
in jprint.h to jprint_util.h which I'm not particularly fond of doing
but it seems worth doing anyway to keep things a bit more organised. Now
the only functions in jprint.c are main(), jprint_sanity_chks() (see
below for why) and usage().

The reason for having the jprint_sanity_chks() function in jprint.c and
not jprint_util.c is because it uses usage() which is the same name as
other tools' usage function so it has to stay in jprint.c. Of course the
function could be renamed jprint_usage() but it feels like it's not so
much a utility anyway and the sanity checks feels like not much of a
utility either but rather a core component of the tool.

There are some functions that should probably remain in jprint.c but
those can be moved back later if that is decided to be the case.