Fix potential NULL pointer dereference in jprint
In the function is_jprint_match() it checked pattern != NULL but it
happily passed pattern->pattern to strcmp(), strcasecmp(), strstr() and
strcasestr(). The pointer should never be NULL but it is theoretically
possible it is.

The function is_jprint_match() now takes another arg, a char *name,
which if NULL is set to pattern->pattern. If both pattern and name are
NULL or name is NULL and pattern->pattern is NULL it is an error. This
simplifies the calls to the above mentioned functions and it also might
prove useful for when no name_arg is specified as the idea of dumping
the entire buffer read in from read_all() is not correct as it means we
can't print the levels etc. Instead we have to traverse the tree and add
the nodes to a single list of matches (not in a pattern) or otherwise
have a function that goes through printing the nodes without worrying
about matches. This is TBD but since the concept of patterns is not
correct (due to an earlier misunderstanding) this might all be changed
anyway.

New version of jprint 0.0.30 2023-06-27.