Partially fix jprint -Y str
A bug in searching for already existing matches was also fixed.

The function add_jprint_match() now takes a char *name_str and char
*value_str and these are used instead of pattern->pattern. The pattern
can now be NULL. If it is then the new jprint->matches list will be
acted on and otherwise pattern->matches list.

Testing of the call to add_jprint_match() where calls are does still
need to be tested more thoroughly (in particular with the name_str and
value_str args) but with the caveat that the value is the same as the
name in the current matches list this cannot be tested as easily.

The purpose of the change to the function is to help with printing the
entire file as there are no patterns (or name_args) when this happens
but until now there was no way to have a match without a pattern (but
for patterns see below). Now a new function or functions can be added so
that goes through the tree but adds every node to the jprint->matches
list.

As the concept of patterns as it stands is not correct but currently is
all that it can be (until the new soon to be added functions are added
to the jparse library) this will do but the struct and list might very
well be removed. However matches are more likely to stay.

This fix to -Y str is not perfect in that in some cases it can end up
matching other types but that depends on the JSON file contents and the
options. It might be said that -Y int is also somewhat improved but
again depending on the json file (h2g2.json which has also been updated)
this cannot be tested entirely.