Add back call to json check tool to jprint
In the process I discovered a use after free that was not detected
before because of the test system it not being a problem but it is in
macOS. As well it was not an actual problem until the tool path and tool
args were added to the struct jprint so it didn't even used to be a
problem.

I also moved the FILE *json_file to the struct jprint as a convenience
though some functions are directly passed it. That could be changed if
so desired.