Improve jprint_match wrt struct json node types
The struct now has a type for the name and a type for the value.
Although in some cases we can only have one type in some we can
determine because a number (for example) cannot be the name so not a
string (if it's a string value that's a number that's another matter).

This is not perfect but it does help in some cases and it will be of
value later on when both nodes can be passed into the function.