Fix adding of value/name matches for num/str
For strings and numbers (but see below) the adding of names versus
values has been corrected. This required a lot of new code and more
modularity is called for but this can be done later. However now one can
get the name or value or both of these types and this goes for the -Y
type option too.

The caveat is that there might be cases that have not been handled as
for each type of name there has to be a check for types of values too
and there might be some I missed for numbers (and there might be the
same problem for strings but I am not sure yet).

Others were not done yet but I wanted this in as it's a
significant accomplishment that will help move forward in the tool. For
that reason I have updated the jprint version again to 0.0.31 2023-06-27.