Add jprint -R option to disable sub-tree searching
New version of jprint - 0.0.32 2023-06-28.

The purpose of this option will be to allow one to find matches that are
not under a hierarchy of name_args. Although it was originally a
misunderstanding it was a misunderstanding also in part because I feel
that there is a great value in having it search in a non recursive way
(in the sense of two name_args as described elsewhere) and it seems like
a waste of the code I already put in. Since there will be no ambiguity
there should not be any problem.

The default will be to search in a recursive way but for now that is
disabled as that feature is not yet implemented.

Fixed a potential memory leak that was discovered at the same time as
well.