Add new json function to get matched node string
New jparse and json parser version 1.0.9 2023-06-28.

By matched node string I refer to the string that matched in the parser
either decoded or not depending on the boolean encoded. This will be
used in jprint to prevent having to use what amounts to almost the same
code repeatedly.

Returns a read-only string. It can return a NULL pointer so it's the
responsibility of the caller to check.

The function is:

    char const *json_get_type_str(struct json *node, bool encoded)

and is in json_util.c.

Updated jparse.3 for the new function.