Add jprint -o and -r options
The -o option is for OR searching. The -r option is to search under
anywhere. Only parsing the option and setting the appropriate booleans
have been added as the concepts are not fleshed out.

Why is -o before -r in the parsing options? This is so that it spells
out or which is what -o does itself! Obviously :-)

With this commit I also sorted the member variables in struct jprint by
category and did the same in alloc_jprint(). I did not do this for
getopt() as I don't want to check repeatedly that I remembered to get
the characters in the string. Okay I did do this for -o and -r but I
mean the rest. :-)