bug_report.sh checks jprint and jprint_test.sh
It now checks that both are executable and assuming all is okay (if not
it's an issue found) it gets the version of the tools. The bug_report.sh
version is now 1.0.2 2023-06-29.

I updated a comment in jprint_test.sh but since no functionality changed
I did not update the version.