New good test JSON files
On 2023-07-02 03:38:24 (UTC-07:00) at 29 km NNE of Point Possession,
Alaska (coordinates 61.176N 150.514W) with a depth of 191.0 km there was
a magnitude of 3.4 earthquake. Although this is not at all related to
the IOCCC the USGS has some JSON files for some of the data and since
both Landon and I are fascinated with volcanoes and earthquakes and
since the files are a different kind I felt that it would be good tests
for jprint later on. The files added start with dyfi_.

Fun fact: Landon did the work to setup that live camera (for Kilauea
which recently erupted) and enable access to thousands of people without
clogging the link between the mainland and the island of Hawaii. I
didn't see any json files for that eruption of last year's Mauna Loa
eruption so I didn't add it.