Add more test json files - good and bad
The good one is a longer file with earthquake data much of which is
similar lines which seems like a good test case for jprint. For this
theme probably no more need to be added but it seems like a fun one and
will be useful (it's also minified which is good for that reason).

The bad one is the same but it's in both bad and bad_loc - with the
proper error file for the latter directory.