Added json-README.md
In an effort to write the specifications for command line utilities
that read JSON, and to assist others in understanding the JSON
related code, code comments, data structures and documentation found
in this GitHub repo, we add a first draft of `jparse/json-README.md`.

We mention this document in the `jparse/README.md` as well.

Updated `CHANGES.md` accordingly.

And of course, comments, corrections and suggestions are welcome!