Fix JSON_NUMBER regexp - resolve #750
New json parser and jparse version 1.0.10 2023-06-29.

A problem was reported on 12 July 2023 where the string 1e0 was reported
as invalid JSON. This is because the regexp in the jparse.l only allowed
(as far as exponent goes) a fraction followed by an exponent rather than
that OR an integer followed by an exponent.

Another couple problems had to be resolved as well here to make make
test work (at the top level directory) okay. First of all a googolplex
is too big so the file googolplex.json was removed from the good
directory. As we don't have a directory for general JSON that is
although valid JSON cannot be parsed this file was removed instead of
moved to another directory. Perhaps we need that functionality in the
test suite too but that can be decided later on (as a reminder we do
have it for the auth.json and info.json files). For reference a
googolplex is
1e10000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000.

The googol.json (yes this is where Google gets their name :-) ) file was
kept as it's not too big for C types.

Finally the test_ioccc/test_JSON/general.json/bad/info.timestamp-0.json
is valid JSON so it was moved to the good subdirectory. I'm not sure how
this was never triggered a problem unless it was just recently added.