Add boolean parsed to json_foo structs
New JSON parser and jparse version: 1.0.11 2023-07-14. Note yesterday
when the version was updated I forgot to update the date. I fixed this
in the CHANGES.md and added a note to the current release in CHANGES.md.

As part of issue #752 I have added to the struct json_ structures a
boolean parsed which if true means that whether or not the values could
be converted is false or true it is still valid JSON.

I discovered and fixed some functions that had missing sanity checks of
setting item->converted to false at first. As well on successful
conversion it appeared that sometimes the item->converted was not set to
true and in the cases I noticed this it has been corrected.

Before an error is issued we now check that BOTH parsed and converted
are false as just because the conversion failed does not mean that it
could not be parsed as valid JSON. Now a googolplex is valid JSON! Added
back the googolplex.json file removed yesterday though note it'll issue
a warning that it cannot be converted!

For numbers this required updating the jnum_{chk,gen,test}.c files as
not doing so would cause make test to fail.

I added placeholder functions for floating notation test and e notation
test which can be added later. It might be a simple matter of checking
the json_process_floating() and adding the specific code to each
function but I am not sure yet.

There are numerous XXX comments that have to be addressed and some
things might need to be changed as well.

There might have been other fixes.

Ran make prep without problems.