Add code for test of floating / e notations
These functions are probably not complete and might not be strictly
correct AND/OR what is needed. Much of the tests do the same as the JSON
checks but these functions have fewer tests as C is less strict in some
things when it comes to floating point than JSON. The JSON parser does
all the checks including the additional checks.

Add check for more than one dot in JSON number strings. Not strictly
necessary as the lexer would report an invalid token but this is for
sanity. This check is also in the new functions.

This does NOT fully resolve issue #752!

Fix issues in jnum test code where '.' was not always detected (that is
the is_floating) and also the e notation (is_e_notation) was not always
detected. Rebuilt jnum_test.c for this to work right.

Running make prep passes.

XXX comments still exist and need to be looked at but again this issue
is not fully resolved and the rest might not be completed by me as the
other parts I am less clear on.