Always set parsed to true if floating / e notation
Previously the boolean was set to true only if it was converted and this
defeats the purpose of the boolean. Well it was set to true in another
place but it's now in the proper place (I did not remove it from the
other places either though as an extra sanity check).