More code for json struct parsed bool
New JSON parser and jparse version 1.0.12 2023-07-15.

Add more checks for parsed bool in structs if converted is false. In the
case of numbers this was already done but for others it was not. However
the checks now check (except for numbers as they're allowed to be
different) that both converted and parsed are true. If only one is true
we print a warning.

Improve how numbers are printed and how strings are printed. For numbers
it's messages improved and typo fixed. For strings there are new flags:
'c' for converted and 'p' for parsed. In the case of strings we do
explicitly check for either or of converted or parsed but it might be
more correct to be like the others: check for both true and warn if only
one of them is true.