Resolve most or all of issue #752
More use of macros added in commit
1e4240bf00fd3abe7e695c0eb302bd682dae12ea.

Renamed sem_node_valid_converted() to sem_node_valid() and update it to
use the macro CONVERTED_PARSED_JSON_NODE (for non-numbers) and
VALID_JSON_NODE for numbers. For numbers either converted or parsed can
be true: if neither are true it is an error. For other types both
converted and parsed have to be true.

Unless I am mistaken this commit fully resolves the issue #752. If this
is so I will then update the jparse and json parser version to 1.1.0 but
in case this is not done yet I haven't yet.