Add parsing of all options in jfmt
Updated usage message.

Added -m max_depth option. This was also added to the
json_util_README.md document.

The option parsing for jfmt should be complete.

Updated version to 0.0.1.

The other tools will probably take more work to get the options parsed
but this one was simple as the code already existed.