Add -o option to common json tool options
Just like with jfmt I believe that the -o ofile option is useful for the
other tools as well and since it's not used by the other tools for
something else I have added it to the json_util_README.md document.

The code for this already exists (as of yesterday) in jfmt (and if
memory serves me right it's complete). It's almost completed in jval but
as I have other things to do with that tool it's not being committed
yet.