Parsing most options of jval done
With the exception of -n and -S which are new concepts all parsing of
options for jval is done. The two remaining options should not take much
work but I wanted this in to get a clean state.

Removed unused code and structs. One function that might have some merit
is in but functionally it does nothing yet.

A bug fix in jfmt is imminent but the rest of jval will probably not
happen today. Once jval options are parsed I will work on jnamval. After
that man pages can be created. I hope that this can be finished by the
end of the week but this might or might not happen depending on a number
of factors including any possible discussion.