Write parsing of jval -S and -n options
This means that all options for jval are now parsed.

Test code is not yet written for these two options but at this time
everything is should be completely parsed. These two options might or
might not have to change when more is decided upon but it seems to be
good at this point.